+++ webpack/src/markdown/deckdeckgo-markdown-plugin.ts
@@ -5,7 +5,7 @@ const path = require('path');
 const readline = require('readline');
 
 const Remarkable = require('remarkable');
-const {replaceEntities, escapeHtml, unescapeMd} = require('remarkable/lib/common/utils');
+const {replaceEntities, escapeHtml, unescapeMd} = require('remarkable');
 
 const util = require('util');
 const readFile = util.promisify(fs.readFile);