+++ docs/src/app/pages/docs/edit/app-edit-reveal/app-edit-reveal.tsx
@@ -1,24 +1,9 @@
-import {Component, Element, h} from '@stencil/core';
-
-import {MenuService} from '../../../../services/menu/menu.service';
+import {Component, h} from '@stencil/core';
 
 @Component({
   tag: 'app-edit-reveal'
 })
 export class AppEditReveal {
-
-  @Element() el: HTMLElement;
-
-  private menuService: MenuService;
-
-  constructor() {
-    this.menuService = MenuService.getInstance();
-  }
-
-  async componentWillLoad() {
-    this.menuService.enable();
-  }
-
   render() {
     return [
       <app-navigation></app-navigation>,