+++ docs/src/app/pages/docs/misc/app-misc-open-source/app-misc-open-source.tsx
@@ -1,24 +1,10 @@
-import {Component, Element, h} from '@stencil/core';
-
-import {MenuService} from '../../../../services/menu/menu.service';
+import {Component, h} from '@stencil/core';
 
 @Component({
   tag: 'app-misc-open-source',
   styleUrl: 'app-misc-open-source.scss'
 })
 export class AppMiscOpenSource {
-
-  @Element() el: HTMLElement;
-
-  private menuService: MenuService;
-
-  constructor() {
-    this.menuService = MenuService.getInstance();
-  }
-
-  async componentWillLoad() {
-    this.menuService.enable();
-  }
 
   render() {
     return [