+++ docs/src/app/pages/docs/slides/app-slide-youtube/app-slide-youtube.tsx
@@ -2,8 +2,6 @@ import {Component, Element, h, Listen} from '@stencil/core';
 
 import {DeckdeckgoDocsUtils} from '../../../../utils/deckdeckgo-docs-utils';
 
-import {MenuService} from '../../../../services/menu/menu.service';
-
 @Component({
   tag: 'app-slide-youtube',
   styleUrl: 'app-slides-youtube.scss'
@@ -11,16 +9,6 @@ import {MenuService} from '../../../../services/menu/menu.service';
 export class AppSlideYoutube {
 
   @Element() el: HTMLElement;
-
-  private menuService: MenuService;
-
-  constructor() {
-    this.menuService = MenuService.getInstance();
-  }
-
-  async componentWillLoad() {
-    this.menuService.enable();
-  }
 
   async componentDidLoad() {
     await DeckdeckgoDocsUtils.reloadCode(this.el);